Better errors when the token is a non-printable control char
I had a heck of a time trying to debug a bit of JSON that had an extra
byte-order mark char in it while testing the previous commit.

This makes the token print out with JSON.stringify, and includes the
char code in hex, so it's easier to tell why it's a problem.